requirements.txt
numpy==1.24.3
torch==2.0.1
transformers==4.30.2
scikit-learn==1.2.2
matplotlib==3.7.1
pandas==2.0.2
tqdm==4.65.0
pytest==7.3.1
pylint==2.17.4
black==23.3.0
networkx==3.1
spacy==3.5.3
nltk==3.8.1
tensorflow==2.12.0
tensorboard==2.12.3
pyyaml==6.0
requests==2.31.0
pymongo==4.4.0
psycopg2-binary==2.9.6
beautifulsoup4==4.12.2
lxml==4.9.2
python-dateutil==2.8.2 